Move test suits setup to one place (#9205)
