+++ wrappers/react/test/reactMemo.spec.tsx
@@ -9,25 +9,6 @@ import {
   sleep,
 } from './_helpers';
 
-const SPEC = {
-  container: null
-};
-
-beforeEach(() => {
-  let container = document.createElement('DIV');
-  container.id = 'hotContainer';
-  document.body.appendChild(container);
-
-  SPEC.container = container;
-});
-
-afterEach(() => {
-  const container = document.querySelector('#hotContainer');
-  container.parentNode.removeChild(container);
-
-  SPEC.container = null;
-});
-
 /**
  * Worth noting, that although it's possible to use React.memo on renderer components, it doesn't do much, as currently they're recreated on every
  * Handsontable's `render`.
@@ -59,7 +40,7 @@ describe('React.memo', () => {
                 }}>
         <MemoizedRendererComponent2 hot-renderer/>
       </HotTable>
-    ), SPEC.container).hotInstance;
+    )).hotInstance;
 
     hotInstance.render();
